Replace client channel `try_send` with `send`
Replace all `try_send` calls on the client channel tx side with a
`send(...).await`. This will wait for a slot in the client channel and
relies on the client task making progress when sending packets. No
packet is silently dropped but a slow client can backpressure on *every*
connected client.