Add send timeout to client task
When a packet cannot be sent withing `client::SINK_SEND_TIMEOUT` the
connection to the client is closed by breaking the tx loop in the client
task. This avoid that a single client can permanently lock the broker
loop trying to enqueue packets in the client channels. The timeout is 1s
and just a random value.