Backpressure on client connections (#27)
* Replace client channel `try_send` with `send`

Replace all `try_send` calls on the client channel tx side with a
`send(...).await`. This will wait for a slot in the client channel and
relies on the client task making progress when sending packets. No
packet is silently dropped but a slow client can backpressure on *every*
connected client.

* Add send timeout to client task

When a packet cannot be sent withing `client::SINK_SEND_TIMEOUT` the
connection to the client is closed by breaking the tx loop in the client
task. This avoid that a single client can permanently lock the broker
loop trying to enqueue packets in the client channels. The timeout is 1s
and just a random value.