Extend and fix subscription identifier encoding and decoding (#46)
* Fix subscription identifier parsing

The subscription identifier is encoded as var int according to
[MQTT-3.8.2.1.2]. Add a parsing test with and without subscription
identifier.

* Add subscription identifier decoding checks MQTT-3.8.2.1.2

Accorindng to MQTT-3.8.2.1.2 the subscription identifier must be set
exactly once and must not be 0. Extend the property parsing for
subscribe packets with those checks.

* Extend subscription identifiers

The Publish packet can have multiple subscription identifiers according
to MQTT-3.3.2.3.8. Subscription identifiers must not have a value 0.